Pulizia del codice

+++ src/shinil35/shat/util/Base64.java
@@ -1046,7 +1046,7 @@ public class Base64
 			{
 				// There's a bad input character in the Base64 stream.
 				throw new java.io.IOException(String.format("Bad Base64 input character decimal %d in array position %d",
-						((int) source[i]) & 0xFF, i));
+						(source[i]) & 0xFF, i));
 			} // end else:
 		} // each input character
 
@@ -1114,7 +1114,7 @@ public class Base64
 		if ((bytes != null) && (bytes.length >= 4) && (!dontGunzip))
 		{
 
-			int head = ((int) bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+			int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
 			if (java.util.zip.GZIPInputStream.GZIP_MAGIC == head)
 			{
 				java.io.ByteArrayInputStream bais = null;
